Configs  - Rename STATUSES to STATES everywhere  - List all services for health
