+++ examples/nocloud_public/docker-compose.yml
@@ -26,6 +26,12 @@ services:
     environment:
       LOG_LEVEL: -1
       SIGNING_KEY: "${SIGNING_KEY}"
+      REGISTRY_HOST: registry:8080
+      SERVICES_HOST: services-registry:8080
+      SP_HOST: sp-registry:8080
+      SETTINGS_HOST: settings:8080
+      DNS_HOST: dns-mgmt:8080
+      STATES_HOST: states:8080
     networks:
       - proxy
       - grpc-internal
@@ -63,6 +69,8 @@ services:
       DB_HOST: db:8529
       DB_CRED: "${DB_USER}:${DB_PASS}"
       DRIVERS: driver-ione:8080
+      SETTINGS_HOST: settings:8080
+      STATES_HOST: states:8080
     depends_on:
       - db
     networks:
@@ -80,6 +88,8 @@ services:
       DB_HOST: db:8529
       DB_CRED: "${DB_USER}:${DB_PASS}"
       DRIVERS: driver-ione:8080
+      SETTINGS_HOST: settings:8080
+      STATES_HOST: states:8080
     networks:
       - proxy
       - arangodb
@@ -118,16 +128,16 @@ services:
     volumes:
       - redis-data:/data
 
-  statuses:
-    image: ghcr.io/slntopp/nocloud/statuses:latest
+  states:
+    image: ghcr.io/slntopp/nocloud/states:latest
     networks:
       - grpc-internal
       - redis
     environment:
       LOG_LEVEL: -1
       PORT: 8080
-      REDIS_HOST: redis_statuses:6379
-  redis_statuses:
+      REDIS_HOST: redis_states:6379
+  redis_states:
     image: redis
     restart: always
     command: redis-server --save 60 1
@@ -136,7 +146,7 @@ services:
     ports:
       - 6379:6379
     volumes:
-      - redis_statuses-data:/data
+      - redis_states-data:/data
 
   driver-ione:
     container_name: driver
@@ -172,5 +182,5 @@ networks:
 
 volumes:
   redis-data:
-  redis_statuses-data:
+  redis_states-data:
   data:
